refactor(react-components): move caching into separate class in architecture (#4910)
* refactor(react-components): move caching into domain object

* chore: preliminary lint fix

* chore: lint fix some more

* chore: remove unwanted changes

* chore: make direct reference from render targe to caches

* fix: tests

* chore: rename function/file
