+++ react-components/src/architecture/base/renderTarget/RevealRenderTarget.ts
@@ -38,6 +38,7 @@ import { Changes } from '../domainObjectsHelpers/Changes';
 import { type CogniteClient } from '@cognite/sdk';
 import { type BaseTool } from '../commands/BaseTool';
 import { ContextMenuController } from './ContextMenuController';
+import { type CdfCaches } from './CdfCaches';
 
 const DIRECTIONAL_LIGHT_NAME = 'DirectionalLight';
 
@@ -50,6 +51,7 @@ export class RevealRenderTarget {
   private readonly _commandsController: CommandsController;
   private readonly _rootDomainObject: RootDomainObject;
   private readonly _contextmenuController: ContextMenuController;
+  private readonly _cdfCaches: CdfCaches;
   private _ambientLight: AmbientLight | undefined;
   private _directionalLight: DirectionalLight | undefined;
   private _clippedBoundingBox: Box3 | undefined;
@@ -65,7 +67,7 @@ export class RevealRenderTarget {
   // CONSTRUCTOR
   // ==================================================
 
-  constructor(viewer: Cognite3DViewer<DataSourceType>, sdk: CogniteClient) {
+  constructor(viewer: Cognite3DViewer<DataSourceType>, sdk: CogniteClient, cdfCaches: CdfCaches) {
     this._viewer = viewer;
 
     const cameraManager = this.cameraManager;
@@ -75,6 +77,7 @@ export class RevealRenderTarget {
     this._commandsController = new CommandsController(this.domElement);
     this._commandsController.addEventListeners();
     this._contextmenuController = new ContextMenuController();
+    this._cdfCaches = cdfCaches;
     this._rootDomainObject = new RootDomainObject(this, sdk);
 
     this.initializeLights();
@@ -117,6 +120,10 @@ export class RevealRenderTarget {
 
   public get contextMenuController(): ContextMenuController {
     return this._contextmenuController;
+  }
+
+  public get cdfCaches(): CdfCaches {
+    return this._cdfCaches;
   }
 
   public get cursor(): string {
