+++ react-components/src/components/CacheProvider/index.ts
@@ -1,13 +1,4 @@
 /*!
  * Copyright 2024 Cognite AS
  */
-
-export type { Image360AnnotationAssetInfo } from './types';
-
-export { useAssetMappedNodesForRevisions } from './AssetMappingAndNode3DCacheProvider';
-export { useFdmAssetMappings } from './NodeCacheProvider';
-export {
-  usePointCloudAnnotationMappingsForModels,
-  usePointCloudAnnotationMappingsForAssetIds
-} from './PointCloudAnnotationCacheProvider';
-export { useImage360AnnotationMappingsForAssetIds } from './Image360AnnotationCacheProvider';
+export { type Image360AnnotationAssetInfo } from './types';
