+++ react-components/src/components/RevealKeepAlive/RevealKeepAlive.tsx
@@ -4,10 +4,6 @@
 
 import { type ReactNode, type ReactElement, useRef, useEffect } from 'react';
 import { RevealKeepAliveContext } from './RevealKeepAliveContext';
-import { type FdmNodeCache } from '../CacheProvider/FdmNodeCache';
-import { type AssetMappingAndNode3DCache } from '../CacheProvider/AssetMappingAndNode3DCache';
-import { type PointCloudAnnotationCache } from '../CacheProvider/PointCloudAnnotationCache';
-import { type Image360AnnotationCache } from '../CacheProvider/Image360AnnotationCache';
 import { type SceneIdentifiers } from '../SceneContainer/sceneTypes';
 import { type RevealRenderTarget } from '../../architecture/base/renderTarget/RevealRenderTarget';
 
@@ -15,10 +11,6 @@ export function RevealKeepAlive({ children }: { children?: ReactNode }): ReactEl
   const renderTargetRef = useRef<RevealRenderTarget>();
   const isRevealContainerMountedRef = useRef<boolean>(false);
   const sceneLoadedRef = useRef<SceneIdentifiers>();
-  const fdmNodeCache = useRef<FdmNodeCache>();
-  const assetMappingCache = useRef<AssetMappingAndNode3DCache>();
-  const pointCloudAnnotationCache = useRef<PointCloudAnnotationCache>();
-  const image360AnnotationCache = useRef<Image360AnnotationCache>();
 
   useEffect(() => {
     return () => {
@@ -31,11 +23,7 @@ export function RevealKeepAlive({ children }: { children?: ReactNode }): ReactEl
       value={{
         renderTargetRef,
         isRevealContainerMountedRef,
-        sceneLoadedRef,
-        fdmNodeCache,
-        assetMappingCache,
-        pointCloudAnnotationCache,
-        image360AnnotationCache
+        sceneLoadedRef
       }}>
       {children}
     </RevealKeepAliveContext.Provider>
