+++ react-components/src/hooks/index.ts
@@ -26,3 +26,7 @@ export type {
   Image360AnnotationMappedAssetData,
   ThreeDModelFdmMappings
 } from './types';
+
+export * from './cad';
+export * from './pointClouds';
+export { useImage360AnnotationMappingsForAssetIds } from './useImage360AnnotationMappingsForAssetIds';
