+++ react-components/src/hooks/pointClouds/index.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright 2024 Cognite AS
+ */
+export { usePointCloudAnnotationMappingForIntersection } from './usePointCloudAnnotationMappingForIntersection';
+export { usePointCloudAnnotationMappingsForAssetIds } from './usePointCloudAnnotationMappingsForAssetIds';
+export { usePointCloudAnnotationMappingsForModels } from './usePointCloudAnnotationMappingsForModels';
