+++ react-components/src/hooks/useImage360AnnotationMappingsForAssetIds.ts
@@ -1,0 +1,49 @@
+/*!
+ * Copyright 2024 Cognite AS
+ */
+
+import { type UseQueryResult, useQuery } from '@tanstack/react-query';
+import {
+  type Image360AnnotationAssetInfo,
+  type Image360AnnotationModel
+} from '../components/CacheProvider/types';
+import { useImage360AnnotationCache } from '../components/CacheProvider/CacheProvider';
+
+export type Image360AnnotationDataResult = {
+  siteId: string;
+  annotationModel: Image360AnnotationModel[];
+};
+
+export const useImage360AnnotationMappingsForAssetIds = (
+  assetIds: Array<string | number> | undefined,
+  siteIds: string[] | undefined
+): UseQueryResult<Image360AnnotationAssetInfo[]> => {
+  const image360AnnotationCache = useImage360AnnotationCache();
+
+  return useQuery({
+    queryKey: [
+      'reveal',
+      'react-components',
+      'image360-annotations-info',
+      ...(assetIds?.map((assetId) => assetId.toString()).sort() ?? []),
+      ...(siteIds?.map((siteId) => siteId).sort() ?? [])
+    ],
+    queryFn: async () => {
+      if (
+        assetIds === undefined ||
+        assetIds.length === 0 ||
+        siteIds === undefined ||
+        siteIds.length === 0
+      ) {
+        return [];
+      }
+      const annotationAssetInfo = await image360AnnotationCache.getReveal360Annotations(siteIds);
+      const filteredAnnotationAssetInfo = annotationAssetInfo.filter((annotationInfo) => {
+        return assetIds.includes(annotationInfo.asset.id);
+      });
+      return filteredAnnotationAssetInfo;
+    },
+    staleTime: Infinity,
+    enabled: assetIds !== undefined && siteIds !== undefined
+  });
+};
