+++ react-components/src/query/use3dNodeByExternalId.tsx
@@ -2,10 +2,10 @@
  * Copyright 2023 Cognite AS
  */
 import { type UseQueryResult, useQuery } from '@tanstack/react-query';
-import { useFdmNodeCache } from '../components/CacheProvider/NodeCacheProvider';
 import { type Node3D } from '@cognite/sdk';
 import { type DmsUniqueIdentifier } from '../data-providers/FdmSDK';
 import { useReveal } from '../components/RevealCanvas/ViewerContext';
+import { useFdmNodeCache } from '../components/CacheProvider/CacheProvider';
 
 export const use3dNodeByExternalId = ({
   externalId,
@@ -30,10 +30,7 @@ export const use3dNodeByExternalId = ({
       }));
 
       const modelMappings = (
-        await fdmNodeCache.cache.getMappingsForFdmInstances(
-          [{ externalId, space }],
-          modelsRevisionIds
-        )
+        await fdmNodeCache.getMappingsForFdmInstances([{ externalId, space }], modelsRevisionIds)
       ).find((model) => model.mappings.size > 0);
 
       const node3d = modelMappings?.mappings.get(externalId)?.[0];
